Remove buffers stat and add quad mesh shaders. Remove drawn buffers label and Add a simple position_texture_color shader for simple primitives like quads to implement static sprite batching in the future.

+++ core/3d/CCBundle3D.h
@@ -113,7 +113,7 @@ public:
                         const char* mtl_basepath = nullptr);
 
     // calculate aabb
-    static AABB calculateAABB(const std::vector<float>& vertex, int stride, const std::vector<unsigned short>& index);
+    static AABB calculateAABB(const std::vector<float>& vertex, int stride, const std::vector<unsigned int>& index);
 
     Bundle3D();
     virtual ~Bundle3D();
